[Commit] Iteration 190 #errors = 26 <= 26.
Addressing `BuildError `[/tmp/tempfspartition/QnetTransformArtifacts/MidTransformCode/sourceCode/MvcAppToModernize/WebApplication/Controllers/ProductController.cs][CS0518][44,55] Predefined type 'System.Int32' is not defined or imported``.

+++ MvcAppToModernize/WebApplication/Controllers/ProductController.cs
@@ -41,7 +41,7 @@ namespace WebApplication.Controllers
             var viewModel = new ProductViewModel
             {
                 Products = products,
-                CartItemCount = cartItems?.Sum(c => c.Quantity) ?? 0
+                CartItemCount = cartItems?.Sum(c => c.Quantity) ?? (System.Int32)0
             };
 
             return Ok(viewModel);
